add credentials
